+++ core-ng/src/test/java/core/framework/impl/web/service/TestWebService.java
@@ -2,7 +2,13 @@ package core.framework.impl.web.service;
 
 import core.framework.api.http.HTTPStatus;
 import core.framework.api.json.Property;
+import core.framework.api.validate.Length;
+import core.framework.api.validate.Max;
+import core.framework.api.validate.Min;
+import core.framework.api.validate.NotEmpty;
 import core.framework.api.validate.NotNull;
+import core.framework.api.validate.Pattern;
+import core.framework.api.validate.Size;
 import core.framework.api.web.service.DELETE;
 import core.framework.api.web.service.GET;
 import core.framework.api.web.service.PUT;
@@ -11,7 +17,11 @@ import core.framework.api.web.service.PathParam;
 import core.framework.api.web.service.QueryParam;
 import core.framework.api.web.service.ResponseStatus;
 
+import java.time.LocalDate;
+import java.time.ZonedDateTime;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 /**
  * @author neo
@@ -23,7 +33,7 @@ public interface TestWebService {
 
     @GET
     @Path("/test/:id")
-    TestResponse get(@PathParam("id") Integer id);
+    Optional<TestResponse> get(@PathParam("id") Integer id);
 
     @PUT
     @Path("/test/:id")
@@ -38,20 +48,62 @@ public interface TestWebService {
     @Path("/test")
     List<TestResponse> batch(List<TestRequest> requests);
 
+    enum TestEnum {
+        @Property(name = "A")
+        A,
+        @Property(name = "B")
+        B
+    }
+
     class TestRequest {
         @NotNull
+        @NotEmpty
+        @Pattern("\\d+.*")
+        @Length(max = 10)
         @Property(name = "string_field")
         public String stringField;
+
+        @Property(name = "items")
+        public List<TestItem> items;
+    }
+
+    class TestItem {
+        @Property(name = "zoned_date_time_field")
+        public ZonedDateTime zonedDateTimeField;
+
+        @Property(name = "enum_field")
+        public TestEnum enumField;
     }
 
     class TestSearchRequest {
         @NotNull
         @QueryParam(name = "int_field")
         public Integer intField;
+
+        @QueryParam(name = "boolean_field")
+        public Boolean booleanField;
+
+        @QueryParam(name = "long_field")
+        public Long longField;
+
+        @Min(1)
+        @Max(100)
+        @QueryParam(name = "double_field")
+        public Long doubleField;
+
+        @QueryParam(name = "date_field")
+        public LocalDate dateField;
     }
 
     class TestResponse {
         @Property(name = "int_field")
         public Integer intField;
+
+        @Size(min = 1)
+        @Property(name = "string_map")
+        public Map<String, String> stringMap;
+
+        @Property(name = "items")
+        public Map<String, TestItem> items;
     }
 }